Beginning the great int purge
So, if we want to use this message code on our embedded boards,
we should use fixed int sizes so that it works as-expected
on all platforms.

+++ common/ExitMessage.hpp
@@ -5,7 +5,7 @@
 class ExitMessage : public Message {
 public:
 
-	ExitMessage(int id) : Message(id) { }
+	ExitMessage(uint16_t id) : Message(id) { }
 
 	static std::unique_ptr<ExitMessage> fromJSON(const Json::Value& object)
 	{
