+++ common/TestMessage.cpp
@@ -12,7 +12,7 @@ const StaticString payloadKey("test payload");
 
 } // end anonymous namespace
 
-TestMessage::TestMessage(int id, const Json::Value& object) :
+TestMessage::TestMessage(uint16_t id, const Json::Value& object) :
 	Message(id),
 	val(object)
 {
